+++ cnccode/controller_classes/CTKPIReport.inc.php
@@ -14,8 +14,8 @@ class CTKPIReport extends CTCNC
     const GET_QUOTATION_CONVERSION                = "quotationConversion";
     const GET_DAILY_STATS                         = "dailyStats";
     const GET_DAILY_SOURCE                        = 'dailySource';
-
     const GET_ENGINEER_MONTHLY_BILLING            = "engineerMonthlyBilling";
+    const GET_DAILY_CONTACT                       = "dailyContact";
     /**
      * CTKPIReport constructor.
      */
@@ -54,6 +54,9 @@ class CTKPIReport extends CTCNC
             case self::GET_DAILY_SOURCE:
                 echo json_encode($this->getDailySource(), JSON_NUMERIC_CHECK);
                 exit;
+            case self::GET_DAILY_CONTACT:
+              echo json_encode($this->getDailyContact(), JSON_NUMERIC_CHECK);
+              exit;
             default:
                 $this->setTemplate();
                 break;
@@ -378,5 +381,28 @@ WHERE problem.`pro_date_raised` >= '2020-01-01'
 
         $query .= "  GROUP BY  invhead.`inh_date_printed_yearmonth`, inl_desc  order by inh_date_printed_yearmonth";
         return DBConnect::fetchAll($query, $params);
+    }
+    function getDailyContact(){
+      $from = (@$_REQUEST['from'] ?? '') == '' ? null : $_REQUEST['from'];;
+      $to         = (@$_REQUEST['to'] ?? '') == '' ? null : $_REQUEST['to'];
+      $where="";
+      $params=[];
+      if($from!=''){      
+        $where .=" AND create_at>=:from";
+        $params["from"]=$from;
+      }
+      if($to!=''){      
+        $where .=" AND create_at<=:to";
+        $params["to"]=$to;
+      }
+      $query="SELECT 'Inbound' AS type ,COUNT(*) total,DATE_FORMAT(create_at,'%Y-%m-%d') date FROM callactivity_customer_contact 
+      WHERE isInbound=1   $where
+      GROUP BY DATE_FORMAT(create_at,'%Y-%m-%d')
+      UNION
+      SELECT 'Outbound' AS type ,COUNT(*) total,DATE_FORMAT(create_at,'%Y-%m-%d') date FROM callactivity_customer_contact 
+      WHERE isInbound=0   $where
+      GROUP BY DATE_FORMAT(create_at,'%Y-%m-%d')";
+      //echo $query; exit;
+      return DBConnect::fetchAll($query, $params);
     }
 }