+++ htdocs/components/shared/DailyStatsComponent/DailyStatsComponent.js
@@ -57,16 +57,18 @@ class DailyStatsComponent extends MainComponent {
                 <td>{this.getOpenSrCard(summary.prioritySummary)}</td>
                 <td>{this.getTeamSrCard(summary.openSrTeamSummary, evenBackgroundColor, eventTextColor)}</td>
                 <td>{this.getDailySourceCard(summary.dailySourceSummary)}</td>
-                <td>{this.getTotalCardWithBiggerNumber("Unique Customers", summary.uniqueCustomerTodaySummary.total, evenBackgroundColor, eventTextColor)}</td>
+                <td>{this.getDailyInboundOutBoundCard(summary.inboundOutbound, evenBackgroundColor, eventTextColor)}</td>                
                 <td>{this.getTotalCardWithBiggerNumber("Near SLA", summary.nearSLASummary.total)}</td>
                 <td>{this.getTotalCardWithBiggerNumber("Near Fix SLA Breach", summary.nearFixSLABreach, evenBackgroundColor, eventTextColor)}</td>
+
             </tr>
-            <tr>
+            <tr>            
                 <td>{this.getTotalCardWithBiggerNumber("Raised Today", summary.raisedTodaySummary.total, evenBackgroundColor, eventTextColor)}</td>
                 <td>{this.getTotalCardWithBiggerNumber("Today's Started", summary.raisedStartTodaySummary.total)}</td>
                 <td>{this.getTotalCardWithBiggerNumber("Fixed Today", summary.fixedTodaySummary.total, evenBackgroundColor, eventTextColor)}</td>
                 <td>{this.getTotalCardWithBiggerNumber("Reopened Today", summary.reopenTodaySummary.total)}</td>
                 <td>{this.getTotalCardWithBiggerNumber("Breached SLA", summary.breachedSLATodaySummary.total, evenBackgroundColor, eventTextColor)}</td>
+                <td>{this.getTotalCardWithBiggerNumber("Unique Customers", summary.uniqueCustomerTodaySummary.total)}</td>
             </tr>
             </tbody>
         </table>
@@ -260,7 +262,24 @@ class DailyStatsComponent extends MainComponent {
             onClick: () => window.open('popup.php?action=dailyStats', 'popup', 'width=1250,height=600')
         })
     }
-
+    getDailyInboundOutBoundCard = (data, backgroundColor = "#C6C6C6", textColor = "#3C3C3C") => {
+        return <div className="sd-card" style={{backgroundColor: backgroundColor, color: textColor}}>
+            <label className="sd-card-title">Daily Contact</label>
+            <table>
+                <tbody>
+                    <tr>
+                        <td>Inbound</td>
+                        <td>{data.inbound}</td>
+                    </tr>
+                    <tr>
+                        <td>Outbound</td>
+                        <td>{data.outbound}</td>
+                    </tr>
+                </tbody>
+            </table>
+        </div>
+       
+    };
     render() {
         return (
             <div className={this.props.className}